feat(toasts): function to create toasts from main
