+++ src/renderer/src/features/Toasts/ToastsContext.tsx
@@ -1,11 +1,5 @@
 import { createContext } from "react";
-
-export interface Toast {
-  message: string;
-  type: "info" | "success" | "danger" | "warning";
-  timeoutMs?: number;
-  noIcon?: boolean;
-}
+import { Toast } from "$shared/types";
 
 export interface InternalToast extends WithRequired<Toast, "timeoutMs"> {
   id: string;
